Added Windows compatibility. Resolves #16, resolves #3
Banana Split on Windows had two issues. The first was due to some difference in the way Windows does something with directory listing. The file type and path would sometimes return false. Exceptions would display in PHP and would ruin the JSON structure that Angular needed.

The second issue involved use of ffmpeg on Windows. Windows has no /dev/null, so NUL was put in its place if the PHP server is running on Windows. Also, ffmpeg on Windows would parse weirdly when detecting black frames, so I put a check in to make sure the parsing wouldn't try to parse invalid data.